+++ src/main/rule/rule-masks.js
@@ -17,6 +17,7 @@ module.exports = (() => ({
     MASK_CONTENT: '$$',
     MASK_CONTENT_EXCEPTION: '$@$',
     MASK_COMMENT: '!',
+    MASK_HOST_FILE_COMMENT: '#',
     MASK_HINT: '!+',
     MASK_DIRECTIVES: '!#',
     MASK_SCRIPTLET: '#%#//scriptlet',